test(core): consolidate reserved provider-data key tests into test_test_context_middleware.py
Per review feedback on #5589, merge the two reserved-key regression tests
into tests/unit/server/test_test_context_middleware.py instead of keeping
them in a separate file. That file already exercises ProviderDataMiddleware
with FastAPI + TestClient, so the new cases fit alongside the existing
__test_id coverage.
